+++ packages/vscode-extension-pack/CHANGELOG.md
@@ -20,4 +20,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ## [0.0.1] - 2026-03-27
 
 ### Added
-- Bundles Anytime Markdown, Anytime Graph, and Anytime Git
+- Bundles Anytime Markdown, Anytime Graph, and Anytime Trail